fix: UserTable componentized
